# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OM-Paie API",
        "description": "API de paiement mobile et transfert d'argent - OM-Paie avec MongoDB",
        "contact": {
            "email": "support@om-paie.sn"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "https://tinkin-transfer.onrender.com/api",
            "description": "Serveur de production sur Render"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/payment": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Effectuer un paiement vers un marchand",
                "description": "Effectue un paiement d'un utilisateur vers un marchand par code marchand",
                "operationId": "payMerchant",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code_marchand",
                                    "montant"
                                ],
                                "properties": {
                                    "code_marchand": {
                                        "description": "Code unique du marchand",
                                        "type": "string",
                                        "example": "BOUT001"
                                    },
                                    "montant": {
                                        "description": "Montant à payer (min: 1, max: 1000000)",
                                        "type": "number",
                                        "format": "float",
                                        "example": 2500
                                    },
                                    "description": {
                                        "description": "Description optionnelle du paiement",
                                        "type": "string",
                                        "example": "Achat produits"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paiement effectué avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Paiement effectué avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "64f7b1a2c5d4e123456789ae"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "payment"
                                                },
                                                "montant": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 2500
                                                },
                                                "expediteur_telephone": {
                                                    "type": "string",
                                                    "example": "+221771234567"
                                                },
                                                "destinataire_telephone": {
                                                    "type": "string",
                                                    "example": "+221338901234"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "completed"
                                                },
                                                "metadata": {
                                                    "properties": {
                                                        "code_marchand": {
                                                            "type": "string",
                                                            "example": "BOUT001"
                                                        },
                                                        "description": {
                                                            "type": "string",
                                                            "example": "Achat produits"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur métier (solde insuffisant, marchand introuvable, etc.)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/payment/check-merchant": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Vérifier un code marchand",
                "description": "Vérifie si un code marchand existe et est actif, retourne les informations du marchand",
                "operationId": "checkMerchant",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code_marchand"
                                ],
                                "properties": {
                                    "code_marchand": {
                                        "description": "Code unique du marchand",
                                        "type": "string",
                                        "example": "BOUT001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Marchand trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "exists": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Marchand trouvé"
                                        },
                                        "merchant": {
                                            "$ref": "#/components/schemas/Merchant"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Marchand non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "exists": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Marchand non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/payment/merchant-info": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Obtenir les informations d'un marchand",
                "description": "Récupère les informations publiques d'un marchand par son code",
                "operationId": "getMerchantInfo",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code_marchand"
                                ],
                                "properties": {
                                    "code_marchand": {
                                        "description": "Code unique du marchand",
                                        "type": "string",
                                        "example": "BOUT001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Informations du marchand récupérées",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "merchant": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "BOUT001"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Boutique Fatou"
                                                },
                                                "telephone": {
                                                    "type": "string",
                                                    "example": "+221775551234"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "fatou@boutique.sn"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "active"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Marchand non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/payment/merchant-history": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Historique des paiements d'un marchand",
                "description": "Récupère l'historique des paiements reçus par un marchand, triés par date décroissante (fonctionnalité réservée aux marchands authentifiés - non implémentée)",
                "operationId": "getMerchantPaymentHistory",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code_marchand"
                                ],
                                "properties": {
                                    "code_marchand": {
                                        "description": "Code unique du marchand",
                                        "type": "string",
                                        "example": "BOUT001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "403": {
                        "description": "Fonctionnalité non disponible",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Fonctionnalité réservée aux marchands authentifiés"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Merchant": {
                "title": "Marchand",
                "description": "Modèle de marchand",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "64f7b1a2c5d4e123456789ae"
                    },
                    "name": {
                        "type": "string",
                        "example": "Boutique Sandaga"
                    },
                    "code": {
                        "type": "string",
                        "example": "SND001"
                    },
                    "telephone": {
                        "type": "string",
                        "example": "+221338901234"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "boutique@sandaga.sn"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "example": "active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "title": "Réponse d'erreur",
                "description": "Réponse standard en cas d'erreur",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Message d'erreur"
                    },
                    "error": {
                        "type": "string",
                        "example": "error_code",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "title": "Erreur de validation",
                "description": "Réponse en cas d'erreur de validation",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "properties": {
                            "field_name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Le champ field_name est requis."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "OAuth2 avec Laravel Passport",
                "flows": {
                    "password": {
                        "tokenUrl": "/oauth/token",
                        "scopes": {}
                    }
                }
            },
            "bearerAuth": {
                "type": "http",
                "description": "Bearer Token (JWT) obtenu après connexion",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}